+++ Darkness/Skill/S_Darkness_Rare_IronMaidensEmbrace.cs
@@ -21,7 +21,7 @@ namespace Darkness
     {
         public override string DescExtended(string desc)
         {
-            return base.DescExtended(desc).Replace("&a", (BChar.GetStat.maxhp * 0.7f).ToString());
+            return base.DescExtended(desc).Replace("&a", ((int)(BChar.GetStat.maxhp * 0.7f)).ToString());
         }
         public override void SkillUseSingle(Skill SkillD, List<BattleChar> Targets)
         {
